Delete nightly build workflows, update workflow names and data classes
- Deleted nightly build workflows
- Updated workflow names to 'package distributable'
- Changed data class properties from Int to Double for top, left, width, height.

+++ src/main/kotlin/nestdrop/XmlDataClasses.kt
@@ -43,13 +43,13 @@ class XmlDataClasses {
         @SerialName("Name")
         val name: String,
         @SerialName("Top")
-        val top: Int,
+        val top: Double,
         @SerialName("Left")
-        val left: Int,
+        val left: Double,
         @SerialName("Width")
-        val width: Int,
+        val width: Double,
         @SerialName("Height")
-        val height: Int,
+        val height: Double,
         @SerialName("Type")
         val type: Int,
         @SerialName("Deck")
